requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
apipkg==1.5
appdirs==1.4.4
Appium-Python-Client==1.0.2
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.1
colorlog==4.6.2
ddt==1.2.1
decorator==4.4.0
distlib==0.3.1
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
execnet==1.8.0
filelock==3.0.12
filetype==1.0.7
Flask==1.1.2
Flask-Cors==3.0.9
har2case==0.3.1
httprunner==2.3.0
idna==2.8
importlib-metadata==0.18
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.11.2
jsonpath==0.82
jsonpath-rw==1.4.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.3
more-itertools==7.0.0
openpyxl==3.0.4
packaging==19.0
pipenv==2020.8.13
pluggy==0.13.1
ply==3.11
prettytable==0.7.2
projects==1.0.9
psutil==5.8.0
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.0
pyparsing==2.4.0
pypiwin32==223
pytest==6.0.1
pytest-forked==1.3.0
pytest-rerunfailures==9.1
pytest-xdist==2.0.0
pytz==2020.1
pywin32==227
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
robotframework==3.1.2
robotframework-pythonlibcore==2.2.1
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.0.0b1
selenium==3.141.0
six==1.12.0
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.24.1
urwid==2.1.2
virtualenv==20.0.35
virtualenv-clone==0.5.4
wcwidth==0.1.7
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
zipp==0.5.1
zope.event==4.5.0
zope.interface==5.1.0
